+++ database/migrations/2019_07_20_164120_create_payment_methods_table.php
@@ -17,7 +17,7 @@ class CreatePaymentMethodsTable extends Migration
             $table->increments('id');
             $table->string('title');
             $table->string('description');
-            $table->integer('price');
+            $table->decimal('price');
             $table->timestamps();
         });
     }
